xmind修改、部分功能提取
