+++ lsc-w91453/src/components/SimpleTable.tsx
@@ -1,5 +1,7 @@
 import { defineComponent, reactive, ref } from "vue";
 import { type TableProps, tableProps } from "./types";
+import { pagination } from './index'
+import './style.css';
 
 interface tableRow {
   id: number;
@@ -34,32 +36,8 @@ export default defineComponent({
         return type.value === 1 ? a[key] - b[key] : b[key] - a[key];	
       }
     }
-
-    // 表格样式
-    const cell = {
-      width: '200px'
-    }
-
-    //分页数据方法
-    const pagelist = (curr: number, total: number) => {
-      let list = [];
-      if(total < 7) {
-        for(var r = 1; r < total + 1; r++) {
-          list.push(r);
-        }
-      } else {
-        if(curr < 4) { 
-          list = [1,2,3,4,'...',total];
-        } else {
-          if(curr > total - 3) {
-            list = [1, '...', total-3,total-2,total-1,total];
-          } else {
-            list = [1,'...',curr-1, curr, curr+1,'...',total];
-          }
-        }
-      }
-      return list;
-    }
+  
+    const { pagelist } = pagination();
     const MAX_NUMBER = ref(10);
     const total = Math.ceil(props.data?.length / MAX_NUMBER.value);
     const currentPage = ref(1);
@@ -71,18 +49,10 @@ export default defineComponent({
       if (n && !isNaN(n)) {
         currentPage.value = n;
         pages.value = pagelist(n, total);
+        console.log(pages)
         list.data = backUpsList.slice((n-1) * MAX_NUMBER.value, n * MAX_NUMBER.value);
       }
     }
-
-    // 分页样式
-    const pagination = {
-        display: 'flex',
-        listStyle:'none'
-    };
-    const liItem = {
-      margin: '0 10px'
-    };
     
     return () => {
       return (
@@ -92,7 +62,7 @@ export default defineComponent({
               <tr>
                   {
                     columns.map(item => {
-                      return <th style={cell} onClick={handleSort.bind(this,item)}>
+                      return <th class="cell" onClick={handleSort.bind(this,item)}>
                         <span>{item}</span>
                         <span id="sort">{Number(type.value) === 0 ? ' 无序' : Number(type.value) === 1 ? ' 正序' : ' 倒序'}</span>
                       </th>
@@ -112,14 +82,14 @@ export default defineComponent({
             </tbody>
           </table>
           <div>
-            <ul style={pagination}>
+            <ul>
                 {
                     pages.value.map(item => {
-                        return <li style={liItem} onClick={pageCtrl.bind(this, item)}>{item}</li>
+                        return <li class="li-item" onClick={pageCtrl.bind(this, item)}>{item}</li>
                     })
                 }
-                <li style={liItem}>跳转至</li>
-                <li style={liItem}>
+                <li class="li-item">跳转至</li>
+                <li class="li-item">
                     <input type="text" v-model={jumpNumber.value} onKeydown={pageCtrl.bind(this, jumpNumber.value)} />
                 </li>
             </ul>
